9term/fipt/requirements/pip.txt
Django==1.10.3
django-registration-redux==1.4
django-braces==1.10.0
py-moneyed==0.6.0
django-money==0.9.1
djangorestframework==3.5.3
Pillow==3.4.2
paypalrestsdk~=1.12.0
django-widget-tweaks==1.4.1
